Update browserify bundle
