+++ package.json
@@ -18,7 +18,6 @@
   },
   "devDependencies": {
     "browserify": "^16.2.2",
-    "nodemon": "^1.17.2",
-    "rm-local-modules": "0.0.2"
+    "nodemon": "^1.17.2"
   }
 }
