+++ views/layout.pug
@@ -10,7 +10,10 @@ html(lang="nl")
 			link(rel="stylesheet" type="text/css" href="/css/style.css")
 
 	body
-		script(src="/bundle.js" async="true" defer="true")
+		script.
+			var streets = JSON.parse(`!{streets}`);
+			var details = JSON.parse(`!{details}`);
+		script(src="/js/bundle.js" async="true" defer="true")
 		script.
 			if ('serviceWorker' in navigator) {
 				navigator.serviceWorker.register('/serviceworker.js')
